Refactor into separate classes
